Refactored Student and StudentHelperClasses
