+++ student/src/main/java/com/bazzaware/Student.java
@@ -1,9 +1,22 @@
 package com.bazzaware;
 
+import java.util.UUID;
+
 public class Student extends StudentHelper {
+    protected String _lastName;
+    protected String _firstName;
+    protected String _course;
+    protected Double _mark = 0.0;
+    protected String _id;
+
+    private Double _minMark = 1.0;
+    private Double _maxMark = 100.0;
 
     public Student(String firstName, String lastName, String course) {
-        super(firstName, lastName, course);
+        _firstName = firstName;
+        _lastName = lastName;
+        _course = course;
+        _id = UUID.randomUUID().toString();
     }
 
     public String getName() {
@@ -25,4 +38,43 @@ public class Student extends StudentHelper {
     public double getMark() {
         return _mark;
     }
+
+    public boolean setMark(String mark) {
+        boolean result = false;
+        if (ConvertStringToDouble(mark)) {
+            Double enteredMark = Double.parseDouble(mark);
+            if (enteredMark >= _minMark && enteredMark <= _maxMark) {
+                _mark = enteredMark;
+                result = true;
+            }
+        } else {
+            System.out.println("Invalid mark, please enter a number between 1 and 100");
+        }
+        return result;
+    }
+
+    public String getGrade() {
+        String grade = "";
+        if (_mark >= 70) {
+            grade = "First";
+        } else if (_mark >= 60) {
+            grade = "2/1";
+        } else if (_mark >= 50) {
+            grade = "2/2";
+        } else if (_mark >= 40) {
+            grade = "Third";
+        } else {
+            grade = "Fail";
+        }
+        return grade;
+    }
+
+    public boolean didPass() {
+        if (_mark >= 40) {
+            return true;
+        } else {
+            return false;
+        }
+    }
+
 }
