+++ student/src/test/java/com/bazzaware/StudentTest.java
@@ -1,6 +1,8 @@
 package com.bazzaware;
 
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 
@@ -28,4 +30,38 @@ public class StudentTest extends Fixtures {
         // assert
         assertNotNull(result);
     }
+    @Test
+    public void ShouldReturnFalseIfLessThanOne() {
+        // arrange
+
+        // act
+        boolean result = student.setMark("0.0");
+
+        // assert
+        assertFalse(result);
+        ;
+    }
+
+    @Test
+    public void ShouldReturnFalseIfMoreThanOneHundred() {
+        // arrange
+
+        // act
+        boolean result = student.setMark("101.0");
+
+        // assert
+        assertFalse(result);
+    }
+
+    public void ShouldReturnTrueIfBetweenOneAndOneHundred() {
+        // arrange
+
+        // act
+        boolean result = student.setMark("1.0");
+
+        // assert
+        assertTrue(result);
+    }
+
+
 }
